# ssa: memory usage optimization around lastDefinitions